python-resume-api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
